added the icon package
formatted code


small fixes to box headers + removed the visible scrollbar from the action sequence box
renamed some files for clarity
