+++ next.config.js
@@ -11,7 +11,7 @@ module.exports = {
   webpack: (config) => {
     config.module.rules.push({
       test: /\.(jpg|ico|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$/i,
-      type: "asset/resource",
+      type: "identifier/resource",
     });
     return config;
   },
