+++ package.json
@@ -7,7 +7,7 @@
     "build": "next build",
     "start": "next start -p $PORT",
     "test": "jest",
-    "test-coverage" : "jest --coverage",
+    "test-coverage": "jest --coverage",
     "export": "next export -o build",
     "deploy": "yarn build && yarn export",
     "swa-start": "swa start build/ --api api",
@@ -20,7 +20,8 @@
     "@emotion/styled": "^11.3.0",
     "@material-ui/core": "^4.11.4",
     "@material-ui/icons": "^4.11.2",
-    "@mui/material": "^5.0.2",
+    "@mui/icons-material": "^5.1.0",
+    "@mui/material": "^5.1.0",
     "axios": "^0.21.1",
     "azure-storage": "^2.10.5",
     "clsx": "^1.1.1",
