+++ src/util/unityContextActions.ts
@@ -1,12 +1,11 @@
 import { UnityContext } from "react-unity-webgl";
 
 const unityContext = new UnityContext({
-    loaderUrl: "build/Build.loader.js",
-    dataUrl: "build/Build.data",
-    frameworkUrl: "build/Build.framework.js",
-    codeUrl: "build/Build.wasm"
+  loaderUrl: "build/Build.loader.js",
+  dataUrl: "build/Build.data",
+  frameworkUrl: "build/Build.framework.js",
+  codeUrl: "build/Build.wasm",
 });
-
 
 function load(arg: string) {
   unityContext.send("Container", "LoadApparatus", arg);
@@ -15,9 +14,9 @@ function voidTrigger(arg: string) {
   unityContext.send("Container", "VoidTrigger", arg);
 }
 
-function callToWebGL(path :string , input:string) {
+function callToWebGL(path: string, input: string) {
   const arg = `${path}@${input}`;
   voidTrigger(arg);
 }
 
-export { callToWebGL, load, voidTrigger ,unityContext};
+export { callToWebGL, load, voidTrigger, unityContext };
